fastapi-project/requirements.txt
alembic==1.18.4
fastapi[standard]===0.135.3
mysql-connector-python==9.6.0
pwdlib==0.3.0
PyJWT==2.12.1
python-dotenv==1.2.2
SQLAlchemy==2.0.49